% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findruns.R
\name{findruns}
\alias{findruns}
\title{Function to find sequences of \code{k} consecutives 1s}
\usage{
findruns(x, k)
}
\arguments{
\item{x}{Vector of 0s and 1s}

\item{k}{Number. Number of desired consecutives 1s}
}
\value{
Vector indicating where the sequences start
}
\description{
Allows you to find to find sequences of \code{k} consecutives 1s in a vector
}
\examples{
y <- c(1, 0, 0, 1, 1, 1, 0, 1, 1)
findruns(y, 2)
findruns(y, 3)
}
