% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_main.R
\name{TL_plot}
\alias{TL_plot}
\title{Plot Method for HDI Class}
\usage{
TL_plot(x, ...)
}
\arguments{
\item{x}{new HDI Class}

\item{...}{Further arguments}
}
\value{
a plot
}
\description{
Plot Method for HDI Class
}
\examples{
data <- TL_read("path/to/hdro_indicators_aggregates_irl.csv")
TL_plot(data)
}
