% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_main.R
\name{TL_print}
\alias{TL_print}
\title{Print Method for HDI Class}
\usage{
TL_print(x, ...)
}
\arguments{
\item{x}{new HDI Class}

\item{...}{Further arguments}
}
\value{
None. Prints the first 10 rows
}
\description{
Print Method for HDI Class
}
\examples{
data <- TL_read("path/to/hdro_indicators_aggregates_irl.csv")
TL_print(data)
}
