% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_main.R
\name{TL_read}
\alias{TL_read}
\title{Read in HDI Data, export a new class with correct title}
\usage{
TL_read(file)
}
\arguments{
\item{file}{file (hdro_indicators_COUNTRYNAME.csv)}
}
\value{
new class "HDI".
}
\description{
Read in HDI Data, export a new class with correct title
}
\examples{
function(file)
}
