% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_main.R
\name{TL_summary}
\alias{TL_summary}
\title{Summary Method for HDI Class}
\usage{
TL_summary(x, ...)
}
\arguments{
\item{x}{new HDI Class}

\item{...}{Further arguments}
}
\value{
summary of the new HDI object.
}
\description{
Summary Method for HDI Class
}
\examples{
data <- TL_read("path/to/hdro_indicators_aggregates_irl.csv")
TL_summary(data)
}
